creates separate file for error classes
