+++ lib/currency_exchange.rb
@@ -1,4 +1,5 @@
 require_relative 'json_lookup'
+require_relative 'error'
 require 'json'
 require 'csv'
 require 'date'
@@ -66,20 +67,3 @@ class CurrencyExchange
   end
 end
 
-class CurrencyError < StandardError
-  def message
-    "Error - unable to calculate rate requested"
-  end
-end
-
-class DateError < StandardError
-  def message
-  "Error - no rate at date found"
-  end
-end
-
-class FileError < StandardError
-  def message
-    "Error - this file type is not accepted. Try JSON or CSV."
-  end
-end
